Suppress Coverity Scan false positives
** CID 646509:       Insecure data handling  (INTEGER_OVERFLOW)
/src/tlslib.c: 156           in cstp_send_file()

** CID 646510:       Insecure data handling  (INTEGER_OVERFLOW)
/src/auth/pam-stack.h: 73           in pam_stack_alloc()

Make sure the suppression annotation:
* appears at the beginning of a C or C++ comment,
* is placed immediately before the line of code where the defect occurs.

+++ src/tlslib.c
@@ -153,6 +153,9 @@ ssize_t cstp_send_file(worker_st *ws, const char *file)
 
 	close(fd);
 
+	/* No way we can read SSIZE_MAX (2 GB on 32-bit systems and 8 EB
+	 * on 64-bit systems) from a profile */
+	// coverity[return_overflow : FALSE]
 	return total;
 }
 
@@ -237,7 +240,7 @@ ssize_t cstp_recv_packet(worker_st *ws, gnutls_datum_t *data, void **p)
 			gnutls_packet_get(packet, data, NULL);
 		}
 	} else {
-		ret = _cstp_recv_packet(ws, ws->buffer, ws->buffer_size);
+		ret = (int)_cstp_recv_packet(ws, ws->buffer, ws->buffer_size);
 		data->data = ws->buffer;
 		data->size = ret;
 	}
